% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{\%+\%}
\alias{\%+\%}
\title{Vectorised addition with NA replacement}
\usage{
x \%+\% y
}
\arguments{
\item{x}{LHS vector to add}

\item{y}{RHS vector to add}

\item{na_replacement}{Replace NA with this value}
}
\description{
Vectorised addition with NA replacement
}

