% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{plus}
\title{NA removal in vectorised addition with multiple inputs (NSE version)}
\usage{
plus(..., na_replacement = 0)
}
\arguments{
\item{...}{vectors to add}

\item{na_replacement}{Replace NA with this value}
}
\description{
NA removal in vectorised addition with multiple inputs (NSE version)
}

