% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{plus_}
\alias{plus_}
\title{NA removal in vectorised addition with multiple inputs}
\usage{
plus_(vecs, na_replacement = 0)
}
\arguments{
\item{vecs}{a list of vectors}

\item{na_replacement}{Replace NA with this value}
}
\description{
NA removal in vectorised addition with multiple inputs
}

